Refactor: split off AnnifRegistry to a separate annif.registry module
