+++ annif/cli.py
@@ -16,6 +16,7 @@ import annif
 import annif.corpus
 import annif.eval
 import annif.project
+import annif.registry
 from annif.project import Access
 from annif.suggestion import SuggestionFilter
 from annif.exception import ConfigurationException, NotSupportedException
@@ -30,7 +31,7 @@ def get_project(project_id):
     """
     Helper function to get a project by ID and bail out if it doesn't exist"""
     try:
-        return annif.project.get_project(project_id, min_access=Access.hidden)
+        return annif.registry.get_project(project_id, min_access=Access.hidden)
     except ValueError:
         click.echo(
             "No projects found with id \'{0}\'.".format(project_id),
@@ -130,7 +131,8 @@ def run_list_projects():
         "Project ID", "Project Name", "Language", "Trained")
     click.echo(header)
     click.echo("-" * len(header))
-    for proj in annif.project.get_projects(min_access=Access.private).values():
+    for proj in annif.registry.get_projects(
+            min_access=Access.private).values():
         click.echo(template.format(
             proj.project_id, proj.name, proj.language, str(proj.is_trained)))
